microservices/user_microservice/app/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
circuitbreaker==2.0.0
click==8.1.7
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
flask-marshmallow==0.15.0
Flask-Caching==2.3.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
logzero==1.7.0
MarkupSafe==2.1.5
marshmallow==3.20.1
packaging==24.0
proofdock-chaos-middleware-flask==0.1.3
proofdock-chaos-middleware-python==0.2.3
python-dotenv==1.0.1
requests==2.31.0
saga-py==1.1.0
tenacity==8.2.3
urllib3==2.2.1
Werkzeug==3.0.2
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
psycopg2-binary==2.9.9
redis==5.0.4
